+++ src/app/core/service/app.service.ts
@@ -3,8 +3,8 @@ import { ElectronProvider } from '@app/provider';
 import { VisibleFlag } from '@app/type/app.type';
 import { IpcRenderer, Remote } from 'electron';
 import { BehaviorSubject, combineLatest, Observable } from 'rxjs';
-import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators';
-import { DialogsService } from './dialogs.service';
+import { debounceTime, map } from 'rxjs/operators';
+import { DialogRefService } from './dialog/dialog-ref.service';
 
 @Injectable({
     providedIn: 'root'
@@ -16,7 +16,7 @@ export class AppService {
 
     constructor(
         private readonly ngZone: NgZone,
-        private readonly dialogs: DialogsService,
+        private readonly dialogRef: DialogRefService,
         electronProvider: ElectronProvider) {
         this.electron = electronProvider.provideRemote();
         this.ipcRenderer = electronProvider.provideIpcRenderer();
@@ -29,7 +29,7 @@ export class AppService {
         this.ipcRenderer.sendSync('register-active-change');
         return combineLatest([
             this.activeChange$,
-            this.dialogs.dialogCountChange()
+            this.dialogRef.dialogCountChange()
         ]).pipe(
             map(([game, dialogCount]) => {
                 let result = VisibleFlag.None;
@@ -46,6 +46,10 @@ export class AppService {
 
     public triggerVisibleChange(): void {
         this.activeChange$.next(this.activeChange$.value);
+    }
+
+    public version(): string {
+        return this.electron.app.getVersion();
     }
 
     public quit(): void {