+++ src/app/core/service/index.ts
@@ -1,10 +1,6 @@
-export * from './input/app.service';
-export * from './input/browser.service';
-export * from './input/clipboard.service';
-export * from './input/keyboard.service';
-export * from './input/mouse.service';
-export * from './input/renderer.service';
-export * from './input/shortcut.service';
-export * from './input/window.service';
-export * from './input/dialogs.service';
+export * from './app.service';
+export * from './browser.service';
+export * from './renderer.service';
+export * from './storage.service';
+export * from './window.service';