+++ src/app/layout/service/user-settings.service.ts
@@ -3,7 +3,7 @@ import { FeatureModule } from '@app/type';
 import { Language } from '@shared/module/poe/type';
 import { Observable, of } from 'rxjs';
 import { flatMap, map } from 'rxjs/operators';
-import { UserSettings } from '../type';
+import { UserSettings, DialogSpawnPosition } from '../type';
 import { UserSettingsDialogService } from './user-settings-dialog.service';
 import { UserSettingsFeatureService } from './user-settings-feature.service';
 import { UserSettingsStorageService } from './user-settings-storage.service';
@@ -29,6 +29,7 @@ export class UserSettingsService {
                     exitAppKeybinding: 'F8',
                     language: Language.English,
                     zoom: 100,
+                    dialogSpawnPosition: DialogSpawnPosition.Cursor,
                     displayVersion: true
                 };