+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.22 (2020-03-01)
 
 - add relaunch button (#275)
+- add dialog spawn position as general setting (Cursor, Center) (#210)
 - fix timeless jewels missing keystone (#274)
 - fix `waterways map of vulnerablity` mismatched with `vulnerablitity skill gem` (#268)